backend/contadores/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
dnspython==2.3.0
frozenlist==1.3.0
gunicorn==20.1.0
idna==3.3
multidict==6.0.2
pymongo==4.3.3
yarl==1.7.2
pytest==7.3.1
pydantic==1.10.9
PyJWT==2.7.0
